requirements.txt
# AI Academic Scheduler - Dependencies

# Core framework
nicegui>=1.4.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter for Supabase

# Authentication & Security
bcrypt>=4.0.0

# Environment variables
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# PDF parsing (for future implementation)
# tabula-py>=2.7.0
# pdfplumber>=0.10.0

# Additional utilities
python-dateutil>=2.8.0

